fix all check errors

+++ src/sources/selfstat.rs
@@ -218,9 +218,9 @@ mod tests {
         let (cpu_time, threads, _, vsize, rss) =
             get_proc_stat("tests/fixtures/proc", 26231).await.unwrap();
 
-        assert_eq!(cpu_time, 17.21);
-        assert_eq!(threads, 1.0);
-        assert_eq!(vsize, 56274944.0);
-        assert_eq!(rss, 1981.0);
+        assert_eq!(cpu_time, 17.21 as f64);
+        assert_eq!(threads, 1.0 as f64);
+        assert_eq!(vsize, 56274944.0 as f64);
+        assert_eq!(rss, 1981.0 as f64);
     }
 }
